fakeroot: split files & libsubid lookups
The shadow utilities and libsubid calls that deal with subid mappings
can now be configured to use a non-files source in `nsswitch.conf`.
This allows e.g. subid mappings to be stored in an LDAP directory and
retrieved via sssd, rather than needing to be defined in
`/etc/sub[ug]id`.

The `nsswitch.conf` configuration is limited to a single source. There
is no explicit or implicit fall back to local files if `subid: sss` is
set. However, the Apptainer implementation of libsubid support still
consults `/etc/sub[ud]id` files with our own code, ahead of retrieving
mappings using libsubid.

This commit modifies the libsubid support picked from Apptainer so that
if Singularity is compiled with libsubid support, it behaves like the
shadow utilities - there is no additional / fallback to local files. If
`nsswitch.conf` is configured to use a source other than `files` then
the local `/etc/sub[ug]id` files are never consulted.

If Singularity is compiled without libsubid support then it will
continues to use its own code to read subid ranges from
`/etc/sub[ug]id` files.

Note that the `config fakeroot` command continues to modify the local
`/etc/sub[ug]id` files in all cases.

To simplifiy the code, and now that there is a route to non-files based
subid mapping via libsubid, the fakeroot id mapping callback for plugins
has been removed.
